V2

+++ photolist.rb
@@ -6,34 +6,30 @@ class PhotoList < Gtk::IconView
     def initialize()
         super()
         @bgthread = nil
-        @dir = nil
+        @db = nil
 
         set_pixbuf_column(1)
         set_text_column(2)
         set_selection_mode(Gtk::SELECTION_MULTIPLE)
         set_columns(0)
-        set_item_width(DEFAULT_IMAGE_WIDTH+50)
+        set_item_width(DEFAULT_IMAGE_WIDTH+10)
 
         # Setup ListStore:
         # Full Path; Thumbnail; Filename
-        @model = Gtk::ListStore.new(String, Gdk::Pixbuf, String)
+        @model = Gtk::ListStore.new(Image, Gdk::Pixbuf, String)
 
         set_model(@model)
+        @pixbufs = []
     end
 
-    def set_directory(dir)
-        @dir = dir
+    def set_database(db)
+        @db = db
 
-        images = Dir.new(dir).to_a.sort!
-
-        $blank = Gdk::Pixbuf.new(Gdk::Pixbuf::COLORSPACE_RGB, false, 8, DEFAULT_IMAGE_WIDTH, DEFAULT_IMAGE_WIDTH*3/4)
-
-        images.each{|im|
-            next if (im.upcase !~ /.*JPG/)
+        db.images.each{|im|
             iter = @model.append
-            @model.set_value(iter, 0, dir + "/" + im)
-            @model.set_value(iter, 1, $blank)
-            @model.set_value(iter, 2, im)
+            @model.set_value(iter, 0, im)
+            @model.set_value(iter, 1, im.thumbnail)
+            @model.set_value(iter, 2, im.filename)
         }
 
         @bgthread = Thread.new {
@@ -44,25 +40,32 @@ class PhotoList < Gtk::IconView
             }
 
             GC.start
-            results = Parallel.map(list) do |filepath|
-                puts "#{path}"
-                pixbuf = Magick::Image::read(filepath).first
-                
-                pix_w = pixbuf.columns
-                pix_h = pixbuf.rows
-                new_h = (pix_h * DEFAULT_IMAGE_WIDTH) / pix_w   # Calculate the scaled height before resizing image
-                out = [filepath, pixbuf.resize(DEFAULT_IMAGE_WIDTH, new_h)]
-                pixbuf.destroy!
-                out
-            end
+            list.each_slice(2){|list_slice|
+                results = Parallel.map(list_slice) do |im|
+                    puts im.inspect
+                    # FIXME: Rotate image if EXIF says so
+                    image = Magick::Image::read(im.fullname).first
+                    
+                    pix_w = image.columns
+                    pix_h = image.rows
+                    r1 = DEFAULT_IMAGE_WIDTH/pix_w.to_f
+                    r2 = DEFAULT_IMAGE_HEIGHT/pix_h.to_f
+                    out = [im.fullname, image.thumbnail([r1, r2].min)]
+                    image.destroy!
+                    out
+                end
 
-            @model.each{|model, path, iter|
-                filepath = @model.get_value(iter, 0)
-                scaled_pix = results.find{|x| x[0] == filepath}[1]
-                pixbuf = Gdk::Pixbuf.new(scaled_pix.export_pixels_to_str(), Gdk::Pixbuf::COLORSPACE_RGB,
-                            false, 8, scaled_pix.columns, scaled_pix.rows, scaled_pix.columns*3)
-                @model.set_value(iter, 1, pixbuf)
-                scaled_pix.destroy!
+                @model.each{|model, path, iter|
+                    im = @model.get_value(iter, 0)
+                    next if (!(result = results.find{|x| x[0] == im.fullname}))
+                    scaled_pix = result[1]
+                    im.thumbnail = Gdk::Pixbuf.new(scaled_pix.export_pixels_to_str(), Gdk::Pixbuf::COLORSPACE_RGB,
+                                false, 8, scaled_pix.columns, scaled_pix.rows, scaled_pix.columns*3)
+                    @model.set_value(iter, 1, im.thumbnail)
+                    scaled_pix.destroy!
+                    # Keep a reference of all the pixbufs (so the memory does not get freed)
+                    #@pixbufs << pixbuf
+                }
             }
         }
     end
